+++ simplecontrols/src/components/autocomplete/autocomplete.js
@@ -1,7 +1,7 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 
-class  Autocomplete extends Component {
+class Autocomplete extends Component {
   constructor(props) {
     super(props);
 
@@ -9,92 +9,94 @@ class  Autocomplete extends Component {
     this.state = {
       resultsExpanded: false,
       searching: false,
-      searchvalue:'',
-      results: []
-    }
+      searchvalue: "",
+      results: [],
+    };
   }
-
 
   render() {
     const onChange = (e) => {
-      if(e.target.value === '')
-      {
+      if (e.target.value === "") {
         this.setState({
           searchvalue: e.target.value,
           searching: false,
-          resultsExpanded:false,
-          results: []
-        })
-      }
-      else
-      {
+          resultsExpanded: false,
+          results: [],
+        });
+      } else {
         this.setState({
           searchvalue: e.target.value,
           searching: true,
-          resultsExpanded:false,
-          results: []
-        })
+          resultsExpanded: false,
+          results: [],
+        });
       }
 
       this.props.onFetchData(e.target.value, (results) => {
         this.setState({
           searching: true,
-          resultsExpanded:true,
-          results : results
-        })
+          resultsExpanded: true,
+          results: results,
+        });
       });
-    }
+    };
 
-    const onSelected =(item) => {
+    const onSelected = (item) => {
       this.props.onSelectedResult(item);
       this.setState({
         searching: false,
-        resultsExpanded:false,
-        searchvalue: item.DisplayText
-      })
-    }
+        resultsExpanded: false,
+        searchvalue: item.DisplayText,
+      });
+    };
 
     const results = () => {
-      if(this.state.resultsExpanded)
-      {
+      if (this.state.resultsExpanded) {
         return (
-        <div className="autocomplete__searchresults">
-          <ul>
-            {
-              this.state.results.map(item => {
-                return <li key={item.DisplayText} onClick={(e) => onSelected(item)}>{item.DisplayText}</li>
-              })
-            }
-          </ul>
-        </div>)
-      }
-      else if(this.state.searching)
-      {
+          <div className="autocomplete__searchresults">
+            <ul>
+              {this.state.results.map((item) => {
+                return (
+                  <li key={item.DisplayText} onClick={() => onSelected(item)}>
+                    {item.DisplayText}
+                  </li>
+                );
+              })}
+            </ul>
+          </div>
+        );
+      } else if (this.state.searching) {
         return (
           <div className="autocomplete__searching">
             <ul>
-              <li><i className="fas fa-cog"/>Searching...</li>
+              <li>
+                <i className="fas fa-cog" />
+                Searching...
+              </li>
             </ul>
-          </div>) 
+          </div>
+        );
       }
-    }
+    };
 
     return (
       <div className="autocomplete" id="autocomplete">
         <div className="autocomplete__searchbox" id="instance-id">
-          <input type="text"  className="autocomplete__input" onChange={onChange} value={this.state.searchvalue} />
-          { 
-            results()
-          }
-          </div>
-
+          <input
+            type="text"
+            className="autocomplete__input"
+            onChange={onChange}
+            value={this.state.searchvalue}
+          />
+          {results()}
+        </div>
       </div>
-    )
+    );
   }
 }
 
 Autocomplete.propTypes = {
   onSelectedResult: PropTypes.func.isRequired,
-  onFetchData: PropTypes.func,  
-}
+  onFetchData: PropTypes.func,
+};
 export default Autocomplete;