+++ simplecontrols/src/index.js
@@ -1,3 +1,3 @@
-import "./styles/main.scss"
-export {default as Autocomplete}  from './components/autocomplete';
-export {default as CheckboxList}  from './components/checkboxlist';
+import "./styles/main.scss";
+export { default as Autocomplete } from "./components/autocomplete";
+export { default as CheckboxList } from "./components/checkboxlist";